refactor : 부서 지역 예외처리 구현
