+++ src/main/java/com/ohgiraffers/hrbank/exception/DuplicatedNameException.java
@@ -1,0 +1,7 @@
+package com.ohgiraffers.hrbank.exception;
+
+public class DuplicatedNameException extends RuntimeException {
+    public DuplicatedNameException() {
+        super("이미 존재하는 이름입니다.");
+    }
+}
